moved jwtsvid file
