+++ examples/spiffe-jwt/server/main.c
@@ -1,11 +1,11 @@
-#include "bundle/jwtbundle/source.h"
-#include "spiffeid/id.h"
-#include "spiffetls/listen.h"
-#include "spiffetls/mode.h"
-#include "spiffetls/option.h"
-#include "svid/jwtsvid/parse.h"
-#include "workload/jwtsource.h"
-#include "workload/x509source.h"
+#include "c-spiffe/bundle/jwtbundle/source.h"
+#include "c-spiffe/spiffeid/id.h"
+#include "c-spiffe/spiffetls/listen.h"
+#include "c-spiffe/spiffetls/mode.h"
+#include "c-spiffe/spiffetls/option.h"
+#include "c-spiffe/svid/jwtsvid/parse.h"
+#include "c-spiffe/workload/jwtsource.h"
+#include "c-spiffe/workload/x509source.h"
 #include <unistd.h>
 
 void init_openssl()
